+++ PokeApiNet.Tests/Cache/ResourceCacheManagerTests.cs
@@ -1,5 +1,4 @@
 ﻿using PokeApiNet.Cache;
-using PokeApiNet.Models;
 using System;
 using Xunit;
 
@@ -130,8 +129,8 @@ namespace PokeApiNet.Tests.Cache
             sut.ClearAll();
 
             // assert
-            Berry retrievedBerry = sut.Get<Berry>(1);
-            Pokedex retrievedPokedex = sut.Get<Pokedex>(1);
+            Berry retrievedBerry = sut.Get<Berry>(berry.Id);
+            Pokedex retrievedPokedex = sut.Get<Pokedex>(pokedex.Id);
             Assert.Null(retrievedBerry);
             Assert.Null(retrievedPokedex);
         }
@@ -151,8 +150,8 @@ namespace PokeApiNet.Tests.Cache
             sut.Clear<Berry>();
 
             // assert
-            Berry retrievedBerry = sut.Get<Berry>(1);
-            Pokedex retrievedPokedex = sut.Get<Pokedex>(1);
+            Berry retrievedBerry = sut.Get<Berry>(berry.Id);
+            Pokedex retrievedPokedex = sut.Get<Pokedex>(pokedex.Id);
             Assert.Null(retrievedBerry);
             Assert.NotNull(retrievedPokedex);
         }