+++ PokeApiNet/Models/Encounters.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace PokeApiNet.Models
+namespace PokeApiNet
 {
     /// <summary>
     /// Methods by which the player might can encounter Pokémon