+++ PokeApiNet/Models/Evolution.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace PokeApiNet.Models
+namespace PokeApiNet
 {
     /// <summary>
     /// Evolution chains are essentially family trees. They start with
@@ -35,6 +35,9 @@ namespace PokeApiNet.Models
         public ChainLink Chain { get; set; }
     }
 
+    /// <summary>
+    /// The linking information between a Pokémon and it's evolution(s)
+    /// </summary>
     public class ChainLink
     {
         /// <summary>
@@ -63,6 +66,9 @@ namespace PokeApiNet.Models
         public List<ChainLink> EvolvesTo { get; set; }
     }
 
+    /// <summary>
+    /// The details for an evolution
+    /// </summary>
     public class EvolutionDetail
     {
         /// <summary>